+++ src/vs/workbench/contrib/userDataSync/electron-sandbox/userDataSync.contribution.ts
@@ -15,8 +15,8 @@ import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { IFileService } from 'vs/platform/files/common/files';
 import { INativeHostService } from 'vs/platform/native/common/native';
-import { INotificationService } from 'vs/platform/notification/common/notification';
-import { CONTEXT_SYNC_STATE, SYNC_TITLE } from 'vs/workbench/services/userDataSync/common/userDataSync';
+import { INotificationService, Severity } from 'vs/platform/notification/common/notification';
+import { CONTEXT_SYNC_STATE, DOWNLOAD_ACTIVITY_ACTION_DESCRIPTOR, IUserDataSyncWorkbenchService, SYNC_TITLE } from 'vs/workbench/services/userDataSync/common/userDataSync';
 import { Schemas } from 'vs/base/common/network';
 
 class UserDataSyncServicesContribution implements IWorkbenchContribution {
@@ -55,6 +55,26 @@ registerAction2(class OpenSyncBackupsFolder extends Action2 {
 			return nativeHostService.showItemInFolder(item.with({ scheme: Schemas.file }).fsPath);
 		} else {
 			notificationService.info(localize('no backups', "Local backups folder does not exist"));
+		}
+	}
+});
+
+registerAction2(class DownloadSyncActivityAction extends Action2 {
+	constructor() {
+		super(DOWNLOAD_ACTIVITY_ACTION_DESCRIPTOR);
+	}
+
+	async run(accessor: ServicesAccessor): Promise<void> {
+		const userDataSyncWorkbenchService = accessor.get(IUserDataSyncWorkbenchService);
+		const notificationService = accessor.get(INotificationService);
+		const hostService = accessor.get(INativeHostService);
+		const folder = await userDataSyncWorkbenchService.downloadSyncActivity();
+		if (folder) {
+			notificationService.prompt(Severity.Info, localize('download sync activity complete', "Successfully downloaded Settings Sync activity."),
+				[{
+					label: localize('open', "Open Folder"),
+					run: () => hostService.showItemInFolder(folder.fsPath)
+				}]);
 		}
 	}
 });
