+++ src/vs/workbench/services/userDataSync/common/userDataSync.ts
@@ -6,12 +6,14 @@
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { IAuthenticationProvider, SyncStatus, SyncResource, IUserDataSyncResource, IResourcePreview } from 'vs/platform/userDataSync/common/userDataSync';
 import { Event } from 'vs/base/common/event';
-import { RawContextKey } from 'vs/platform/contextkey/common/contextkey';
+import { ContextKeyExpr, RawContextKey } from 'vs/platform/contextkey/common/contextkey';
 import { localize } from 'vs/nls';
 import { URI } from 'vs/base/common/uri';
 import { Codicon } from 'vs/base/common/codicons';
 import { registerIcon } from 'vs/platform/theme/common/iconRegistry';
 import { IView } from 'vs/workbench/common/views';
+import { Categories } from 'vs/platform/action/common/actionCommonCategories';
+import { IAction2Options } from 'vs/platform/actions/common/actions';
 
 export interface IUserDataSyncAccount {
 	readonly authenticationProviderId: string;
@@ -45,7 +47,7 @@ export interface IUserDataSyncWorkbenchService {
 	accept(resource: IUserDataSyncResource, conflictResource: URI, content: string | null | undefined, apply: boolean): Promise<void>;
 
 	getAllLogResources(): Promise<URI[]>;
-	downloadSyncActivity(location: URI): Promise<void>;
+	downloadSyncActivity(): Promise<URI | undefined>;
 }
 
 export function getSyncAreaLabel(source: SyncResource): string {
@@ -90,3 +92,11 @@ export const SHOW_SYNC_LOG_COMMAND_ID = 'workbench.userDataSync.actions.showLog'
 // VIEWS
 export const SYNC_VIEW_CONTAINER_ID = 'workbench.view.sync';
 export const SYNC_CONFLICTS_VIEW_ID = 'workbench.views.sync.conflicts';
+
+export const DOWNLOAD_ACTIVITY_ACTION_DESCRIPTOR: Readonly<IAction2Options> = {
+	id: 'workbench.userDataSync.actions.downloadSyncActivity',
+	title: { original: 'Download Settings Sync Activity', value: localize('download sync activity title', "Download Settings Sync Activity") },
+	category: Categories.Developer,
+	f1: true,
+	precondition: ContextKeyExpr.and(CONTEXT_ACCOUNT_STATE.isEqualTo(AccountStatus.Available), CONTEXT_SYNC_STATE.notEqualsTo(SyncStatus.Uninitialized))
+};
